refactor: utility functions
