+++ src/utilities/get-unit-group.ts
@@ -1,0 +1,6 @@
+import { UnitGroup } from '../constants/enums.js';
+
+const getUnitGroup = (unitGroupValue: string): UnitGroup =>
+  Object.values(UnitGroup).find((unitGroup) => unitGroup === unitGroupValue) ?? UnitGroup.US;
+
+export default getUnitGroup;
